extract logic into new classes and unwonkify
